Remove old comment
It's not nonsense. It just doesn't do anything that I can tell.
`f.read()` returns a string. then `bytes()` makes that string into a
bytes object, specifying the encoding.

`etree.XML()` is perfectly happy to take a string or a bytes object,
And the code produces the same output when tested on levanto_catechismo
with and without the bytes function here.

There's a chance that `bytes(..., encoding="utf-8")` is
performing some useful function here, most likely related to the
encoding.  Maybe catching some weird characters or something? So I'm
reluctant to remove it.

Resources:

- https://lxml.de/parsing.html#serialising-to-unicode-strings
- https://stackoverflow.com/a/16037481/8970630
- https://www.codesdope.com/blog/article/bytes-in-python/