Optimize conversion
I implemented basic memoization of the generated HTML, then writing both 
a normal and preview output from it.

This change assumes that every version we want to generate will have a 
corresponding preview that we also want to generate along with it.

I also cleaned up outline and preview outline generation code to be more 
readable